Fixed several compiler warnings. Most fixes involved adding the "const" keyword.
git-svn-id: file:///home/hardaker/lib/sf-bkups/net-snmp-convert-svnrepo/trunk@18889 06827809-a52a-0410-b366-d66718629ded

+++ net-snmp/agent/mibgroup/mibII/interfaces.c
@@ -2746,7 +2746,7 @@ var_ifEntry(struct variable * vp,
         return (u_char *) & long_return;
     case NETSNMP_IFSPEED:
         if (if_ptr)
-            long_return = if_ptr->speed;
+            long_return = (long) if_ptr->speed;
         else
         long_return = (long) ifRow.dwSpeed;
         return (u_char *) & long_return;
